+++ db/migrate/20240523112726_make_transfer_form_m_task_optional.rb
@@ -1,0 +1,5 @@
+class MakeTransferFormMTaskOptional < ActiveRecord::Migration[7.0]
+  def change
+    add_column :transfer_tasks_data, :form_m_not_applicable, :boolean
+  end
+end